+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * httpClient: support to trust specific https cert
 * http: removed undertow HTTP metrics 
     undertow is using jboss threading and EnhancedQueueExecutor, there is overhead to collect stats with jmx, and they are not critical stats 
+* redis: start preparing for Redis6/RESP3 (later version may requires redis 6)
+* httpClient: update okHTTP to 4.6.0 
 
 ### 7.4.1.1 (4/24/2020)
 * http: improve x-forwarded-for parser to support rfc7239, as Azure AG will use ipv4:port format in x-forwarded-for header